requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
idna==3.3
numpy==1.22.0
opencv-python==4.5.5.62
Pillow==9.0.0
python-dotenv==0.19.2
requests==2.27.0
sqlparse==0.4.2
urllib3==1.26.7
